[DEV-847] Maps: UX Feedback
