+++ packages/map/src/CdcMap.js
@@ -10,7 +10,6 @@ import ResizeObserver from 'resize-observer-polyfill'
 import ReactTooltip from 'react-tooltip'
 import chroma from 'chroma-js'
 import parse from 'html-react-parser'
-import Papa from 'papaparse'
 
 // Data
 import colorPalettes from '../../core/data/colorPalettes'
@@ -121,29 +120,6 @@ const CdcMap = ({ className, config, navigationHandler: customNavigationHandler,
 
   let legendMemo = useRef(new Map())
   let innerContainerRef = useRef()
-
-  const DownloadButton = memo(() => {
-    let rawData = state.data || {}
-    const fileName = `${state.general.title || 'data-table'}.csv`
-
-    const csvData = Papa.unparse(rawData)
-
-    const blob = new Blob([csvData], { type: 'text/csv;charset=utf-8;' })
-
-    const saveBlob = () => {
-      //@ts-ignore
-      if (typeof window.navigator.msSaveBlob === 'function') {
-        //@ts-ignore
-        navigator.msSaveBlob(blob, fileName)
-      }
-    }
-
-    return (
-      <a download={fileName} type='button' onClick={saveBlob} href={URL.createObjectURL(blob)} aria-label='Download this data in a CSV file format.' className={`${state.headerColor} btn btn-download no-border`} id={`${skipId}`} data-html2canvas-ignore role='button'>
-        Download Data (CSV)
-      </a>
-    )
-  }, [state.data])
 
   useEffect(() => {
     try {
@@ -1496,8 +1472,6 @@ const CdcMap = ({ className, config, navigationHandler: customNavigationHandler,
             <section className='download-buttons'>
               {state.general.showDownloadImgButton && <MediaDownloadButton text='Download Image' title='Download Chart as Image' type='image' state={config} elementToCapture={imageId} />}
               {state.general.showDownloadPdfButton && <MediaDownloadButton text='Download PDF' title='Download Chart as PDF' type='pdf' state={config} elementToCapture={imageId} />}
-              {/* {config.table.download && <DownloadButton data={stateData ?? {}} />} */}
-              {state.general.showDownloadPdfButton === true && <DownloadButton />}
             </section>
 
             {state.runtime.editorErrorMessage.length === 0 && true === dataTable.forceDisplay && general.type !== 'navigation' && false === loading && (
