+++ packages/map/src/scss/datatable.scss
@@ -1,6 +1,6 @@
 .data-table-container {
   margin: 0;
   @include breakpointClass(md) {
-    margin: 1em;
+    margin: 5px 1em;
   }
 }
